SUI->MUI migration: #1333-#1341 complete, remove SUI deps + CSS

+++ src/components/rangeUsePlanPage/additionalRequirements/index.tsx
@@ -1,14 +1,14 @@
-import React, { useState } from 'react';
+import React from 'react';
 import AdditionalRequirementRow from './AdditionalRequirementRow';
-import { InfoTip } from '../../common';
+import { InfoTip, PrimaryButton, MuiIcon } from '../../common';
 import { IfEditable } from '../../common/PermissionsField';
 import { ADDITIONAL_REQUIREMENTS } from '../../../constants/fields';
 import * as strings from '../../../constants/strings';
 import { FieldArray } from 'formik';
-import { Button, Confirm } from 'semantic-ui-react';
 import uuid from 'uuid-v4';
 import { deleteAdditionalRequirement } from '../../../api';
 import { resetAdditionalRequirementId } from '../../../utils/helper/additionalRequirement';
+import useConfirm from '../../../providers/ConfrimationModalProvider';
 
 interface AdditionalRequirementsProps {
   additionalRequirements: any[];
@@ -16,7 +16,7 @@ interface AdditionalRequirementsProps {
 }
 
 function AdditionalRequirements({ additionalRequirements, planId: _planId }: AdditionalRequirementsProps) {
-  const [indexToRemove, setIndexToRemove] = useState<number | null>(null);
+  const confirm = useConfirm()!;
 
   return (
     <FieldArray
@@ -30,10 +30,10 @@ function AdditionalRequirements({ additionalRequirements, planId: _planId }: Add
               <InfoTip header={strings.ADDITIONAL_REQUIREMENTS} content={strings.ADDITIONAL_REQUIREMENTS_TIP} />
             </div>
             <IfEditable permission={ADDITIONAL_REQUIREMENTS.CATEGORY}>
-              <Button
+              <PrimaryButton
                 type="button"
-                basic
-                primary
+                inverted
+                compact
                 onClick={() => {
                   push({
                     id: uuid(),
@@ -44,9 +44,9 @@ function AdditionalRequirements({ additionalRequirements, planId: _planId }: Add
                 }}
                 className="icon labeled rup__pastures__add-button"
               >
-                <i className="add circle icon" />
+                <MuiIcon name="add circle" />
                 Add Requirement
-              </Button>
+              </PrimaryButton>
             </IfEditable>
           </div>
           <div className="rup__divider" />
@@ -62,32 +62,24 @@ function AdditionalRequirements({ additionalRequirements, planId: _planId }: Add
                 <AdditionalRequirementRow
                   key={additionalRequirement.id}
                   additionalRequirement={additionalRequirement}
-                  onDelete={() => setIndexToRemove(i)}
+                  onDelete={async () => {
+                    const choice = await confirm({
+                      titleText: 'Delete additional requirement',
+                      contentText: 'Are you sure?',
+                    });
+                    if (!choice) return;
+                    const requirement = additionalRequirements[i];
+                    if (!uuid.isUUID(requirement.id)) {
+                      deleteAdditionalRequirement(requirement.planId, requirement.id);
+                    }
+                    remove(i);
+                  }}
                   onCopy={() => push(resetAdditionalRequirementId(additionalRequirements[i]))}
                   namespace={`additionalRequirements.${i}`}
                 />
               ))
             )}
           </div>
-
-          <Confirm
-            header="Delete additional requirement"
-            content="Are you sure?"
-            open={indexToRemove !== null}
-            onCancel={() => {
-              setIndexToRemove(null);
-            }}
-            onConfirm={async () => {
-              const requirement = additionalRequirements[indexToRemove!];
-
-              if (!uuid.isUUID(requirement.id)) {
-                deleteAdditionalRequirement(requirement.planId, requirement.id);
-              }
-
-              remove(indexToRemove!);
-              setIndexToRemove(null);
-            }}
-          />
         </div>
       )}
     />
